Add test for controller view
