+++ tests/test_views.py
@@ -3,22 +3,11 @@ from uuid import uuid4
 
 import pytest
 from django.urls import reverse
-from pytest_django.asserts import assertContains, assertRedirects, assertTemplateUsed
+from pytest_django.asserts import assertContains, assertTemplateUsed
 
 from main.views import ProcessAction
 
-
-class LoginRequiredTest:
-    """Tests for views that require authentication."""
-
-    endpoint: str
-
-    def test_login_redirect(self, client):
-        """Test that the view redirects to the login page."""
-        response = client.get(self.endpoint)
-        assert response.status_code == HTTPStatus.FOUND
-
-        assertRedirects(response, reverse("main:login") + f"?next={self.endpoint}")
+from .utils import LoginRequiredTest
 
 
 class TestIndexView(LoginRequiredTest):
